Add files via upload
Trying to get download_d1_data.R to work for non-eml metadata. We determined that Lines 86-92 in the download_d1_data.R file is generally where the function starts to fail when the input is a data_url to data in this repo (https://search.dataone.org/view/10.24431/rw1k45w)/ any repo that uses non-eml metadata. For this example I used the data_url for the file named 'nga_TGX201809_seabird_processed_densities_L1.csv'. This code that shows the method that produces the most complete eml object so far (from non-eml metadata).